backend/requirements.txt
flask>=3.1.2
flask-sqlalchemy>=3.1.1
flask-session>=0.5.0
werkzeug>=3.1.3
python-dotenv>=1.1.1
psycopg2-binary>=2.9.10
flask-cors>=6.0.1
gunicorn>=23.0.0
openai>=1.0.0
